Improved startup behavior. Improved robustness of network connection attempts. Added "breathing" animation while establishing network connection.

+++ cancel_startup.sh
@@ -1,4 +1,7 @@
 #!/bin/bash
 DIRNAME="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null && pwd )"
 PIDFILENAME="startup.pid"
-kill -SIGINT `cat "${DIRNAME}/${PIDFILENAME}"` && rm "${DIRNAME}/${PIDFILENAME}"
+
+if [ -f "${DIRNAME}/${PIDFILENAME}" ]; then
+	kill -SIGINT `cat "${DIRNAME}/${PIDFILENAME}"` && rm "${DIRNAME}/${PIDFILENAME}"
+fi
